rename PressReleasesContainer to PressReleaseCollection, remove class from HorizontalRule
