--- src/components/press-releases/PressReleasesContainer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./PressReleasesContainer";
